+++ backend/canisters/user_index/api/src/updates/c2c_report_message.rs
@@ -10,6 +10,8 @@ pub struct Args {
     pub message: Message,
     pub already_deleted: bool,
     pub is_public: bool,
+    #[serde(default)]
+    pub csam: bool,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]